+++ database/provider.go
@@ -6,12 +6,11 @@ import (
 
 	dbtypes "github.com/forbole/bdjuno/v3/database/types"
 	dbutils "github.com/forbole/bdjuno/v3/database/utils"
-
-	providertypes "github.com/ovrclk/akash/x/provider/types/v1beta2"
+	"github.com/forbole/bdjuno/v3/types"
 )
 
 // SaveProviders allows to store the providers inside the database
-func (db *Db) SaveProviders(providers []providertypes.Provider, height int64) error {
+func (db *Db) SaveProviders(providers []*types.Provider, height int64) error {
 	if len(providers) == 0 {
 		return nil
 	}
@@ -35,10 +34,12 @@ func (db *Db) SaveProviders(providers []providertypes.Provider, height int64) er
 }
 
 // saveProviders allows to store providers inside the database
-func (db *Db) saveProviders(paramsNumber int, providers []providertypes.Provider, height int64) error {
-	stmt := `INSERT INTO provider (owner_address, host_uri, attributes, info, height) VALUES `
+func (db *Db) saveProviders(paramsNumber int, providers []*types.Provider, height int64) error {
+
+	stmt := `INSERT INTO akash_provider (owner_address, host_uri, attributes, info, height) VALUES `
 	var params []interface{}
 
+	accounts := make([]types.Account, len(providers))
 	for i, provider := range providers {
 
 		bi := i * paramsNumber
@@ -55,7 +56,8 @@ func (db *Db) saveProviders(paramsNumber int, providers []providertypes.Provider
 			return fmt.Errorf("error while converting provider info to DbProviderInfo: %s", err)
 		}
 
-		params = append(params, provider.Owner, provider.HostURI, string(attributesBz), infoValue, height)
+		params = append(params, provider.OwnerAddress, provider.HostURI, string(attributesBz), infoValue, height)
+		accounts[i] = types.NewAccount(provider.OwnerAddress)
 	}
 
 	stmt = stmt[:len(stmt)-1]
@@ -65,9 +67,14 @@ ON CONFLICT ON CONSTRAINT unique_provider DO UPDATE
 		attributes = excluded.attributes,
 		info = excluded.info,
 	    height = excluded.height 
-WHERE provider.height <= excluded.height`
+WHERE akash_provider.height <= excluded.height`
 
-	_, err := db.Sql.Exec(stmt, params...)
+	err := db.SaveAccounts(accounts)
+	if err != nil {
+		return fmt.Errorf("error while storing accounts: %s", err)
+	}
+
+	_, err = db.Sql.Exec(stmt, params...)
 	if err != nil {
 		return fmt.Errorf("error while storing providers: %s", err)
 	}
@@ -81,6 +88,103 @@ func (db *Db) DeleteProvider(ownerAddress string) error {
 	_, err := db.Sql.Exec(stmt, ownerAddress)
 	if err != nil {
 		return fmt.Errorf("error while deleting provider: %s", err)
+	}
+	return nil
+}
+
+// GetAkashProviders returns the akash provider addresses stored inside the database
+func (db *Db) GetAkashProviders() ([]string, error) {
+	stmt := `SELECT owner_address FROM akash_provider`
+
+	var addresses []string
+	if err := db.Sqlx.Select(&addresses, stmt); err != nil {
+		return nil, err
+	}
+
+	if len(addresses) == 0 {
+		return nil, fmt.Errorf("no provider was saved")
+	}
+
+	return addresses, nil
+}
+
+// GetProviderHostURI returns the akash provider's host URI stored inside the database
+func (db *Db) GetProviderHostURI(address string) (string, error) {
+	stmt := `SELECT host_uri FROM akash_provider where owner_address = $1`
+
+	var hostURI []string
+	if err := db.Sqlx.Select(&hostURI, stmt, address); err != nil {
+		return "", err
+	}
+
+	if len(hostURI) == 0 {
+		return "", fmt.Errorf("no host uri is found for provider address %s", address)
+	}
+
+	return hostURI[0], nil
+}
+
+// SaveProviderInventoryStatus allows to store provider inventory status inside the database
+func (db *Db) SaveProviderInventoryStatus(status *types.ProviderInventoryStatus) error {
+	stmt := `INSERT INTO akash_provider_inventory 
+	( 
+		provider_address, active, lease_count, 
+		bidengine_order_count, manifest_deployment_count, 
+		cluster_public_hostname, inventory_status_raw, 
+		active_inventory_sum, pending_inventory_sum, available_inventory_sum, 
+		height 
+	) VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11) 
+ON CONFLICT (provider_address) DO UPDATE 
+	SET active = excluded.active, 
+		lease_count = excluded.lease_count, 
+		bidengine_order_count = excluded.bidengine_order_count, 
+		manifest_deployment_count = excluded.manifest_deployment_count, 
+		cluster_public_hostname = excluded.cluster_public_hostname, 
+		inventory_status_raw = excluded.inventory_status_raw, 
+		active_inventory_sum = excluded.active_inventory_sum, 
+		pending_inventory_sum = excluded.pending_inventory_sum, 
+		available_inventory_sum = excluded.available_inventory_sum, 
+		height = excluded.height 
+	WHERE akash_provider_inventory.height <= excluded.height`
+
+	// Marshal inventory raw status
+	inventoryStatusBz, err := json.Marshal(&status.InventoryStatusRaw)
+	if err != nil {
+		return fmt.Errorf("error while marshaling provider inventory raw status: %s", err)
+	}
+
+	// Get value for active inventory sum
+	active := dbtypes.NewDbAkashResource(status.ActiveInventorySum)
+	activeValue, _ := active.Value()
+
+	// Get value for pending inventory sum
+	pending := dbtypes.NewDbAkashResource(status.PendingInventorySum)
+	pendingValue, _ := pending.Value()
+
+	// Get value for available inventory sum
+	available := dbtypes.NewDbAkashResource(status.AvailableInventorySum)
+	availableValue, _ := available.Value()
+
+	_, err = db.Sql.Exec(stmt,
+		status.ProviderAddress, status.Active, status.LeaseCount,
+		status.BidengineOrderCount, status.ManifestDeploymentCount,
+		status.ClusterPublicHostname, string(inventoryStatusBz),
+		activeValue, pendingValue, availableValue,
+		status.Height,
+	)
+	if err != nil {
+		return fmt.Errorf("error while storing inventory status of provider %s: %s", status.ProviderAddress, err)
+	}
+
+	return nil
+}
+
+// SetProviderStatus allows to update the provider status inside the database
+func (db *Db) SetProviderStatus(providerAddress string, active bool, height int64) error {
+	stmt := `UPDATE akash_provider_inventory SET active = $1 WHERE provider_address = $2`
+	_, err := db.Sql.Exec(stmt, active, providerAddress)
+	if err != nil {
+		return fmt.Errorf("error while updating active status of provider %s: %s", providerAddress, err)
 	}
 	return nil
 }