+++ CHANGELOG.md
@@ -6,10 +6,11 @@
   - `x/deployment`: store params from genesis
   - `x/market`: store params from genesis
   - `x/provider`: store providers data
-
-#### custom modules
 - ([\#453](https://github.com/forbole/bdjuno/pull/453))
   - `x/market`: store active leases
+- ([\#457](https://github.com/forbole/bdjuno/pull/457))
+  - `x/provider`: store provider inventory status
+
 
 ## Version v3.2.0
 ### Changes